requirements.txt
Django==4.2.6
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2-binary==2.9.8
python-decouple==3.5
python-environ==0.4.54
dj-database-url==2.1.0
whitenoise==6.6.0
